+++ src/main/java/com/first/bulletinboard/domain/dto/comment/CommentCreateResponse.java
@@ -11,10 +11,10 @@ import java.time.LocalDateTime;
 @Builder
 @Getter
 public class CommentCreateResponse {
-    private Integer id;
+    private Long id;
     private String comment;
     private String userName;
-    private Integer postId;
+    private Long postId;
     private LocalDateTime createdAt;
     private LocalDateTime lastModifiedAt;
